refactor: simplify questions
