+++ src/Components/Questions/CountryQuestion/CountryQuestion.jsx
@@ -1,7 +1,20 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import countries from '../../../../tasks_configs/capital/countries';
+import { getRandom } from '../../Helpers/utils';
 import QuestionWithTextAnswer from '../BaseQuestions/QuestionWithTextAnswer/QuestionWithTextAnswer';
+import { addQuestion } from '../questionTypes';
 
+
+function generate() {
+  const random = getRandom(0, countries.length);
+  const item = countries[random];
+  return {
+    type: 'country',
+    capital: item.capital,
+    solution: item.country,
+  };
+}
 
 function CountryQuestion({ capital, ...baseComponentProps }) {
   const hint = 'Which country\'s capital is this city:';
@@ -14,4 +27,5 @@ CountryQuestion.propTypes = {
   capital: PropTypes.string.isRequired,
 };
 
-export default CountryQuestion;
+addQuestion('country', { render: CountryQuestion, generate });
+export default generate;
