+++ src/Components/Questions/CountryQuestion/CountryQuestion.test.js
@@ -1,0 +1,10 @@
+import generate from './CountryQuestion';
+
+describe('generateCountryTask', () => {
+  it('should have right types', () => {
+    const task = generate();
+    expect(task.type).toBe('country');
+    expect(task.capital).toBeString();
+    expect(task.solution).toBeString();
+  });
+});
