+++ src/Components/Questions/GreaterLessOrEqualQuestion/GreaterLessOrEqualQuestion.jsx
@@ -1,6 +1,27 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import { getRandom } from '../../Helpers/utils';
 import QuestionWithButtons from '../BaseQuestions/QuestionWithButtons/QuestionWithButtons';
+import { addQuestion } from '../questionTypes';
+
+function getSolution(left, right) {
+  if (left < right) return '<';
+  if (left > right) return '>';
+  return '=';
+}
+
+function generate() {
+  const operands = [
+    getRandom(10, 100),
+    getRandom(10, 100),
+  ];
+  const solution = getSolution(...operands);
+  return {
+    type: 'greaterLessOrEqual',
+    operands,
+    solution,
+  };
+}
 
 const answers = [
   {
@@ -31,4 +52,5 @@ GreaterLessOrEqualQuestion.propTypes = {
   operands: PropTypes.arrayOf(PropTypes.number).isRequired,
 };
 
-export default GreaterLessOrEqualQuestion;
+addQuestion('greaterLessOrEqual', { render: GreaterLessOrEqualQuestion, generate });
+export { generate, getSolution };
