+++ src/Components/Questions/ListeningQuestion/ListeningQuestion.test.js
@@ -1,0 +1,10 @@
+import generate from './ListeningQuestion';
+
+describe('generateListeningTask', () => {
+  it('should have right types', () => {
+    const task = generate();
+    expect(task.type).toBe('listening');
+    expect(task.name).toBeString();
+    expect(task.solution).toBeString();
+  });
+});
