+++ src/Components/Questions/LogoQuestion/LogoQuestion.test.js
@@ -1,0 +1,9 @@
+import { generate } from './LogoQuestion';
+
+describe('generateLogoTask', () => {
+  it('should have right types', () => {
+    const task = generate();
+    expect(task.type).toBe('logo');
+    expect(task.name).toBeString();
+  });
+});
