+++ src/Components/Questions/MathQuestion/MathQuestion.jsx
@@ -1,7 +1,23 @@
 import PropTypes from 'prop-types';
 import React from 'react';
+import { calculateMathSolution, getRandom, getRandomItemFromArray } from '../../Helpers/utils';
 import QuestionWithTextAnswer from '../BaseQuestions/QuestionWithTextAnswer/QuestionWithTextAnswer';
+import { addQuestion } from '../questionTypes';
 
+function generate() {
+  const operands = [
+    getRandom(10, 100),
+    getRandom(10, 100),
+  ];
+  const sign = getRandomItemFromArray(['+', '-']);
+  const solution = calculateMathSolution({ operands, sign });
+  return {
+    type: 'math',
+    operands,
+    sign,
+    solution,
+  };
+}
 
 function MathQuestion({ operands, sign, ...baseComponentProps }) {
   return (
@@ -17,5 +33,5 @@ MathQuestion.propTypes = {
   operands: PropTypes.arrayOf(PropTypes.number).isRequired,
   sign: PropTypes.string.isRequired,
 };
-
-export default MathQuestion;
+addQuestion('math', { render: MathQuestion, generate });
+export default generate;
