+++ src/Components/Questions/MissedWordQuestion/MissedWordQuestion.test.js
@@ -1,4 +1,4 @@
-import { createShortcutCode } from './MissedWordQuestion';
+import { createShortcutCode, generate } from './MissedWordQuestion';
 
 describe('createShortcutCode', () => {
   it('should create shortcut code for 0 based index', () => {
@@ -13,5 +13,17 @@ describe('createShortcutCode', () => {
 
   it('should throw error for indexes bigger than 8', () => {
     expect(() => createShortcutCode(9)).toThrow(/9/);
+  });
+});
+
+describe('generateMissedWordTask', () => {
+  it('should have right types', () => {
+    const task = generate();
+    expect(task.type).toBe('missedWord');
+    expect(task.parts).toBeArrayOfSize(2);
+    expect(task.parts[0]).toBeString();
+    expect(task.suggestions).toBeArray();
+    expect(task.suggestions[0]).toBeString();
+    expect(task.solution).toBeString();
   });
 });
