+++ src/Components/Questions/NumberToStringQuestion/NumberToStringQuestion.test.js
@@ -1,0 +1,9 @@
+import { generate } from './NumberToStringQuestion';
+
+describe('generateNumberToStringTask', () => {
+  it('should have right types', () => {
+    const task = generate();
+    expect(task.type).toBe('numberToString');
+    expect(task.number).toBeNumber();
+  });
+});
