+++ src/Components/Questions/TrueFalseQuestion/TrueFalseQuestion.jsx
@@ -1,6 +1,19 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import trueFalseQuestions from '../../../../tasks_configs/true_false/questions';
+import { getRandom } from '../../Helpers/utils';
 import QuestionWithButtons from '../BaseQuestions/QuestionWithButtons/QuestionWithButtons';
+import { addQuestion } from '../questionTypes';
+
+function generate() {
+  const random = getRandom(0, trueFalseQuestions.length);
+  const item = trueFalseQuestions[random];
+  return {
+    type: 'trueFalse',
+    question: item.question,
+    solution: String(item.right),
+  };
+}
 
 const answers = [
   {
@@ -29,4 +42,5 @@ TrueFalseQuestion.propTypes = {
   question: PropTypes.string.isRequired,
 };
 
-export default TrueFalseQuestion;
+addQuestion('trueFalse', { render: TrueFalseQuestion, generate });
+export default generate;
